Refactor: Split HomeView into smaller components for better readability
