--- UnivApp/General/View/Search/SearchViewModelType.swift
@@ -1,14 +1,0 @@
-//
-//  SearchViewModelType.swift
-//  UnivApp
-//
-//  Created by 정성윤 on 9/26/24.
-//
-
-import Foundation
-
-enum SearchViewModelType: Hashable {
-    case play
-    case list
-    //TODO: - 뷰모델 추가
-}
