+++ UnivApp/View/Category/Play/PlaySegment/SchoolSegmentView.swift
@@ -49,8 +49,8 @@ struct SchoolSegmentView: View {
                     .font(.system(size: 18, weight: .bold))
                     .padding(.leading, 20)
                 
-                SearchView(isFocused: self._isFocused, searchText: $listViewModel.searchText, color: .white)
-                    .environmentObject(self.listViewModel)
+//                SearchView(isFocused: self._isFocused, searchText: $listViewModel.searchText, color: .white)
+//                    .environmentObject(self.listViewModel)
                 
                 ForEach(listViewModel.summaryArray, id: \.self) { item in
                     PlayViewCell(playViewModel: self.viewModel, summaryModel: item)
